getting something working

+++ src/hr_driver.c
@@ -4,10 +4,12 @@
 #include "gpio_driver.h"
 #include "utilities.h"
 
+// Function should be called as many times as possible. In the future, a timer interrupt
+// could be used to run this code.
 cur_heart_rate_t get_hr_periodic(void)
 {
     static uint8_t last_val = 10;
-    static uint8_t last_millis = 0;
+    static uint32_t last_millis = 0;
     static cur_heart_rate_t hr;
 
     // Rudimentary peak detection using a digital input and assuming
